requirements.txt
openai
tiktoken
iso639-lang
flask
argparse

